Fixed README section on tests and did some more testing in test_records
